users_search
-added feature to search through users by username or email
-got rid of friend button if not logged in
